requirements.txt
python-dotenv
openai
azure-identity
rich